쿠폰 기본적인 로직 전부 작성

+++ back/lacheln/model/src/main/java/aba3/lucid/domain/coupon/dto/CouponRequest.java
@@ -25,7 +25,7 @@ public class CouponRequest {
 
     // 쿠폰 할인율
     @Min(5) @Max(90)
-    private BigInteger discountRate;
+    private Integer discountRate;
 
     // 쿠폰 가격 하한선
     @Min(0)
